# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: RanchTrack Cattle Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      summary: List farm tasks
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, completed]
      responses:
        "200":
          description: Task list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Post a new job/task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskInput"
      responses:
        "201":
          description: Task created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{id}:
    patch:
      operationId: updateTask
      tags: [tasks]
      summary: Update task (including mark complete)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskUpdate"
      responses:
        "200":
          description: Updated task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Delete task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /fields:
    get:
      operationId: listFields
      tags: [fields]
      summary: List all fields/pastures
      responses:
        "200":
          description: Field list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Field"
    post:
      operationId: createField
      tags: [fields]
      summary: Add a new field/pasture
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FieldInput"
      responses:
        "201":
          description: Field created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Field"
  /fields/{id}:
    patch:
      operationId: updateField
      tags: [fields]
      summary: Update field
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FieldUpdate"
      responses:
        "200":
          description: Updated field
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Field"
    delete:
      operationId: deleteField
      tags: [fields]
      summary: Delete field
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /employees:
    get:
      operationId: listEmployees
      tags: [employees]
      summary: List all staff members
      responses:
        "200":
          description: Staff list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Employee"
    post:
      operationId: createEmployee
      tags: [employees]
      summary: Add a staff member
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmployeeInput"
      responses:
        "201":
          description: Employee added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
  /employees/{id}:
    patch:
      operationId: updateEmployee
      tags: [employees]
      summary: Update employee
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmployeeUpdate"
      responses:
        "200":
          description: Updated employee
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
    delete:
      operationId: deleteEmployee
      tags: [employees]
      summary: Delete employee
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /settings:
    get:
      operationId: getSettings
      tags: [settings]
      summary: Get farm settings
      responses:
        "200":
          description: Farm settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FarmSettings"
    patch:
      operationId: updateSettings
      tags: [settings]
      summary: Update farm settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SettingsUpdate"
      responses:
        "200":
          description: Updated settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FarmSettings"
  /settings/account:
    delete:
      operationId: deleteAccount
      tags: [settings]
      summary: Delete farm account and all data
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccountDelete"
      responses:
        "200":
          description: Account deleted
components:
  schemas:
...
    Task:
      type: object
      required: [id, title, status, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        assignedToId:
          type: ["integer", "null"]
        assignedToName:
          type: ["string", "null"]
        dueDate:
          type: ["string", "null"]
        timeToFinish:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, completed]
        completedAt:
          type: ["string", "null"]
        createdAt:
          type: string
    TaskInput:
      type: object
      required: [title]
      properties:
        title:
          type: string
        description:
          type: string
        assignedToId:
          type: ["integer", "null"]
        dueDate:
          type: string
        timeToFinish:
          type: string
    TaskUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: ["string", "null"]
        assignedToId:
          type: ["integer", "null"]
        dueDate:
          type: ["string", "null"]
        timeToFinish:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, completed]
    Field:
      type: object
      required: [id, name, status, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        area:
          type: ["number", "null"]
        status:
          type: string
          enum: [available, occupied, resting]
        boundary:
          type: ["string", "null"]
        latitude:
          type: ["number", "null"]
        longitude:
          type: ["number", "null"]
        color:
          type: string
        createdAt:
          type: string
    FieldInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string
        area:
          type: number
        status:
          type: string
          enum: [available, occupied, resting]
        boundary:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        color:
          type: string
    FieldUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        area:
          type: ["number", "null"]
        status:
          type: string
          enum: [available, occupied, resting]
        boundary:
          type: ["string", "null"]
        latitude:
          type: ["number", "null"]
        longitude:
          type: ["number", "null"]
        color:
          type: string
    Employee:
      type: object
      required: [id, fullName, username, role, createdAt]
      properties:
        id:
          type: integer
        fullName:
          type: string
        username:
          type: string
        role:
          type: string
          enum: [employer, employee]
        position:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        createdAt:
          type: string
    EmployeeInput:
      type: object
      required: [fullName, username, password, role]
      properties:
        fullName:
          type: string
        username:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [employer, employee]
        position:
          type: string
        phone:
          type: string
    EmployeeUpdate:
      type: object
      properties:
        fullName:
          type: string
        role:
          type: string
          enum: [employer, employee]
        position:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
    FarmSettings:
      type: object
      required: [farmId, farmName, ownerName, email]
      properties:
        farmId:
          type: integer
        farmName:
          type: string
        ownerName:
          type: string
        email:
          type: string
        location:
          type: ["string", "null"]
    SettingsUpdate:
      type: object
      properties:
        farmName:
          type: string
        ownerName:
          type: string
        location:
          type: ["string", "null"]
    AccountDelete:
      type: object
      required: [password]
      properties:
        password:
          type: string