# lib/api-zod/src/generated/types/fieldUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * RanchTrack Cattle Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { FieldUpdateStatus } from './fieldUpdateStatus';

export interface FieldUpdate {
  name?: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  area?: number | null;
  status?: FieldUpdateStatus;
  /** @nullable */
  boundary?: string | null;
  /** @nullable */
  latitude?: number | null;
  /** @nullable */
  longitude?: number | null;
  color?: string;
}